# Conflicts:
#	src/main/java/jm/task/core/jdbc/dao/UserDaoJDBCImpl.java

+++ src/main/java/jm/task/core/jdbc/model/User.java
@@ -52,7 +52,7 @@ public class User {
         this.lastName = lastName;
     }
 
-    public int getAge() {
+    public byte getAge() {
         return age;
     }
